+++ inc/scoreJSON.d.ts
@@ -1,0 +1,19 @@
+
+declare class SheetDocument {}
+declare class PitchContextTable {}
+
+
+interface NoteLinking {
+	ids: string[];
+	staffTrack: number;
+	contextIndex: number;
+}
+
+
+interface ScoreJSON {
+	doc: SheetDocument;
+	hashTable: {[key: string]: any};
+	midi: object;
+	noteLinkings: NoteLinking[];
+	pitchContextGroup: PitchContextTable[];
+}