+++ tests/bundleScore.ts
@@ -3,13 +3,16 @@ import fs from "fs";
 
 import "../env.js";
 import {markScore} from "../backend/scoreMaker";
+import LogRecorder from "../inc/logRecorder";
 
 
 
 const main = async (...args) => {
+	const logger = new LogRecorder({enabled: true});
+
 	const ly = fs.readFileSync(args[0]).toString();
-	const result = await markScore(ly);
-	console.log("result:", result);
+	const result = await markScore(ly, {logger});
+	console.log("result:", result, logger);
 
 	const outputPath = args[1];
 	if (outputPath) {