+++ tools/batchScoreMaker.ts
@@ -6,21 +6,16 @@ import "../env.js";
 import * as ScoreMaker from "../backend/scoreMaker";
 import walkDir from "../backend/walkDir";
 import asyncCall from "../inc/asyncCall";
+import LogRecorder from "../inc/logRecorder";
 
 
 
 const main = async () => {
 	//console.log("argv:", argv);
 
-	/*const xml = fs.readFileSync(args[0]).toString();
-	const markup = args[1] && fs.readFileSync(args[1]).toString();
-	
-	const ly = await ScoreMaker.xmlToLyWithMarkup(xml, undefined, markup);
-	console.log("result:", ly);*/
-
 	const markup = argv.markup ? fs.readFileSync(argv.markup).toString() : null;
 
-	const lilyFiles = [];
+	const lilyFiles: Set<string> = new Set();
 
 	if (argv.inputXmlDir) {
 		const counting = {
@@ -28,6 +23,7 @@ const main = async () => {
 			failure: 0,
 		};
 
+		// TODO: set xml options according to argv
 		const xmlOptions = {};
 
 		const xmlFiles = walkDir(argv.inputXmlDir, /\.xml$/, {recursive: true});
@@ -39,17 +35,17 @@ const main = async () => {
 			try {
 				const xml = fs.readFileSync(xmlPath).toString();
 				const ly = await ScoreMaker.xmlToLyWithMarkup(xml, xmlOptions, markup);
-	
-				if (!argv.noLyWrite) {
-					asyncCall(fs.writeFile, lyPath, ly).then(() => console.log(lyPath, "wrote."));
-					lilyFiles.push(lyPath);
+
+				if (!argv.noLyWrite && !argv.noWrite) {
+					asyncCall(fs.writeFile, lyPath, ly).then(() => console.log("wrote:", lyPath));
+					lilyFiles.add(lyPath);
 				}
 
 				++counting.success;
 			}
 			catch (err) {
 				console.error(err);
-				console.warn("Error when converting", xmlPath);
+				console.warn("Error when converting xml", xmlPath);
 
 				++counting.failure;
 			}
@@ -58,7 +54,67 @@ const main = async () => {
 		console.log("XML to ly finished, success:", counting.success, "failure:", counting.failure);
 	}
 
-	// TODO:
+	if (argv.inputLyDir) {
+		const lyFiles = walkDir(argv.inputLyDir, /\.ly$/, {recursive: true});
+		lyFiles.forEach(path => lilyFiles.add(path));
+	}
+
+	// TODO: load midi files
+
+	if (argv.inputLyDir || argv.bundleScore) {
+		const counting = {
+			success: 0,
+			perfect: 0,
+			failure: 0,
+		};
+
+		const issues = [];
+
+		for (const lyPath of lilyFiles) {
+			//console.log("lyPath:", lyPath);
+
+			const scorePath = lyPath.replace(/\.\w+$/, ".score.json");
+
+			try {
+				const logger = new LogRecorder({enabled: true});
+
+				const ly = fs.readFileSync(lyPath).toString();
+				const score = await ScoreMaker.markScore(ly, {logger});
+
+				const matchStat = logger.records.reverse().find(record => record.desc === "markScore.match");
+				console.assert(matchStat, "No matchStat in logger.");
+
+				if (matchStat.data.coverage < 1) {
+					console.log("imperfect matching:", lyPath, matchStat.data);
+
+					issues.push({
+						lyPath,
+						coverage: matchStat.data.coverage,
+						omitC: matchStat.data.omitC,
+						omitS: matchStat.data.omitS,
+					});
+				}
+				else
+					++counting.perfect;
+
+				if (!argv.noWrite)
+					asyncCall(fs.writeFile, scorePath, JSON.stringify(score)).then(() => console.log("wrote:", scorePath));
+
+				++counting.success;
+			}
+			catch (err) {
+				console.error(err);
+				console.warn("Error when making score", scorePath);
+
+				++counting.failure;
+			}
+		}
+
+		console.log("Score making finished, perfect:", counting.perfect, ", success:", counting.success, "failure:", counting.failure);
+
+		issues.sort((i1, i2) => i1.coverage - i2.coverage);
+		console.log("issues:", issues);
+	}
 };